add category input validation

+++ server/controllers/categoryControllers.js
@@ -1,63 +1,101 @@
 const Product = require('../models/Product')
 const { Category } = require('./../models/category')
 
-module.exports.get_all_categories = async (req, res) => {
+const getAllCategories = async (req, res) => {
   try {
     const categories = await Category.find()
-    res.status(200).json(categories)
+    res.status(200).json({ categoriesNo: categories.length, categories })
   } catch (err) {
     console.log(err)
     return res.status(500).json({ message: 'Internal server error' })
   }
 }
 
-module.exports.get_category_by_id = async (req, res) => {
+const getCategoryById = async (req, res) => {
   const id = req.params.id
   try {
-    const category = await Category.find()
-    const productsInCategory = await Product.find({ categoryId: category._id })
-    res.status(200).json({ category, productsInCategory })
+    const category = await Category.findById(id)
+    if (category) {
+      const productsInCategory = await Product.find({ categoryId: id })
+      res.status(200).json({ category, productsInCategory })
+    } else {
+      res.status(404).send({ message: 'Category not found' })
+    }
   } catch (err) {
     console.log(err)
     return res.status(500).json({ message: 'Internal server error' })
   }
 }
 
-module.exports.add_category = async (req, res) => {
+const addCategory = async (req, res) => {
   const { name, description } = req.body
   try {
     const category = await Category.create({ name, description })
-    res.status(201).json({ message: 'category added successfully', category })
+    res.status(201).json({ message: 'Category added successfully', category })
   } catch (err) {
     console.log(err)
     return res.status(500).json({ message: 'Internal server error' })
   }
 }
 
-module.exports.delete_category = async (req, res) => {
+const deleteCategory = async (req, res) => {
   const id = req.params.id
   try {
-    const ackCategory = await Category.deleteOne({ _id: id })
-    console.log(ackCategory)
-    res.status(204).json({ message: 'category deleted successfully' })
+    const category = await Category.findOne({ _id: id })
+    if (category) {
+      const result = await Category.deleteOne({ _id: id })
+      if (!result.acknowledged) {
+        return res.status(500).json({ message: 'Internal server error' })
+      }
+      res.status(204)
+    } else {
+      res.status(404).send({ message: 'Category not found' })
+    }
   } catch (err) {
     console.log(err)
     return res.status(500).json({ message: 'Internal server error' })
   }
 }
 
-module.exports.update_category = async (req, res) => {
+const updateCategory = async (req, res) => {
   const id = req.params.id
-  const { name, description } = req.body
+  const updatedCategoryFields = req.body
   try {
-    const ackCategory = await Category.updateOne(
-      { _id: id },
-      { name, description }
-    )
-    console.log(ackCategory)
-    res.status(200).json({ message: 'category updated successfully' })
+    Object.keys(updatedCategoryFields).forEach((key) => {
+      if (
+        updatedCategoryFields[key] === '' ||
+        updatedCategoryFields[key] === undefined
+      ) {
+        delete updatedCategoryFields[key]
+      }
+    })
+    const category = await Category.findOne({ _id: id })
+    if (category) {
+      const result = await Category.updateOne(
+        { _id: id },
+        updatedCategoryFields
+      )
+      if (!result.acknowledged) {
+        return res.status(500).json({ message: 'Internal server error' })
+      }
+      const updatedCategory = await Category.findById(id)
+
+      res
+        .status(200)
+        .json({ message: 'Category updated successfully', updatedCategory })
+    } else {
+      res.status(404).send({ message: 'Category not found' })
+    }
   } catch (err) {
     console.log(err)
     return res.status(500).json({ message: 'Internal server error' })
   }
+}
+
+module.exports = {
+  getAllCategories,
+  getCategoryById,
+  addCategory,
+  updateCategory,
+  deleteCategory,
 }
